updated chatProvider and related components
