+++ lib/providers/chat.provider.dart
@@ -1,17 +1,29 @@
 import 'package:flutter/foundation.dart';
-import 'package:pusher_client/pusher_client.dart';
 import 'package:rada_egerton/config.dart';
+import 'package:rada_egerton/constants.dart';
+import 'package:rada_egerton/utils/main.dart';
+import 'package:pusher_client/pusher_client.dart';
+import 'package:rada_egerton/entities/ChatDto.dart';
 import 'package:rada_egerton/services/auth/main.dart';
-import 'package:rada_egerton/services/constants.dart';
-import 'package:rada_egerton/services/utils.dart';
+import 'package:rada_egerton/entities/UserChatsDTO.dart';
+import 'package:rada_egerton/services/counseling/main.dart';
 
-class ChatProvider extends ChangeNotifier {
-  late PusherClient _pusher;
+class ChatProvider with ChangeNotifier {
   late Channel channel;
-  String _channelName = "radaComms";
   late String _userId;
+  late PusherClient _pusher;
+
+  List<ChatPayload> _privateMsgs = [];
+  List<Msg> _groupMsgs = [];
+  List<Msg> _forumMsgs = [];
+
+  String _channelName = "radaComms";
+
+  CounselingServiceProvider _service = CounselingServiceProvider();
+
   ChatProvider() {
     init();
+    getConversations();
   }
   void init() async {
     String? _autoken = await ServiceUtility.getAuthToken();
@@ -29,10 +41,66 @@ class ChatProvider extends ChangeNotifier {
     });
   }
 
+  List<ChatPayload> get privateMessages {
+    return [...this._privateMsgs];
+  }
+
+  List<Msg> get groupMessages {
+    return [...this._groupMsgs];
+  }
+
+  List<Msg> get forumMessages {
+    return [...this._forumMsgs];
+  }
+
+  void getConversations() async {
+    var results = await _service.fetchUserMsgs();
+    results!.fold(
+      (userChats) {
+        this._privateMsgs =
+            userChats.data.payload.peerMsgs.map(convertToChatPayload).toList();
+        this._groupMsgs = userChats.data.payload.groupMsgs;
+        this._forumMsgs = userChats.data.payload.forumMsgs;
+      },
+      (error) => print('Error from fetchChats() :${error.message}'),
+    );
+    notifyListeners();
+  }
+
+  ChatPayload convertToChatPayload(PeerMsg msg) => ChatPayload(
+      id: msg.id,
+      message: msg.message,
+      imageUrl: msg.imageUrl,
+      senderId: msg.senderId,
+      groupsId: msg.groupsId,
+      reply: msg.reply,
+      status: msg.status,
+      reciepient: msg.reciepient);
+
   @override
   void dispose() {
     // TODO: implement dispose
     super.dispose();
     channel.unbind(ChatEvent.CHAT);
+  }
+
+  sendPeerCounselingMessage(ChatPayload chat, String userId) async {
+    var service = CounselingServiceProvider();
+    final result = await service.peerCounseling(chat, userId);
+    result!.fold((chat) {
+      this._privateMsgs.add(
+            ChatPayload(
+              id: chat.data.payload.id,
+              message: chat.data.payload.message,
+              imageUrl: chat.data.payload.imageUrl,
+              senderId: chat.data.payload.senderId,
+              groupsId: chat.data.payload.groupsId ?? "",
+              reply: chat.data.payload.reply ?? "",
+              status: chat.data.payload.status,
+              reciepient: chat.data.payload.reciepient,
+            ),
+          );
+    }, (r) => null);
+    notifyListeners();
   }
 }
