+++ lib/screens/CounselingTabs/CounselorsTab.dart
@@ -1,14 +1,14 @@
-import 'package:cached_network_image/cached_network_image.dart';
+import '../../sizeConfig.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:rada_egerton/entities/ChatDto.dart';
-import 'package:rada_egerton/entities/UserChatsDTO.dart';
-import 'package:rada_egerton/services/constants.dart';
-import 'package:rada_egerton/services/counseling/main.dart';
-import 'package:rada_egerton/widgets/ChatsScreen.dart';
-import 'package:rada_egerton/widgets/ratingBar.dart';
+import 'package:rada_egerton/constants.dart';
 import '../../providers/counselors.provider.dart';
-import '../../sizeConfig.dart';
+import 'package:rada_egerton/entities/ChatDto.dart';
+import 'package:rada_egerton/widgets/ratingBar.dart';
+import 'package:rada_egerton/widgets/ChatsScreen.dart';
+import 'package:rada_egerton/entities/UserChatsDTO.dart';
+import 'package:rada_egerton/services/counseling/main.dart';
+import 'package:cached_network_image/cached_network_image.dart';
 
 class CounselorsTab extends StatelessWidget {
   const CounselorsTab({Key? key}) : super(key: key);
